backend/requirements.txt
aiosmtpd==1.4.6
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
atpublic==7.0.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
dnspython==2.8.0
fastapi==0.123.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imaplib2==3.6
jiter==0.12.0
lxml==6.0.2
mailtrap==2.4.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mysql-connector-python==9.5.0
openai==2.9.0
premailer==3.10.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
pypdf==6.4.0
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
secure-smtplib==0.1.1
sendgrid==6.12.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
Werkzeug==3.1.4
yagmail==0.15.293
